02_captioning/requirements.txt
transformers==4.39.0
torch==2.1.2 
torchvision==0.16.2
Pillow
opencv-python-headless
huggingface_hub
accelerate
